wiki/4_md_to_openai_lance/requirements.txt
pylance>=0.20
pyarrow
